+++ spec/requests/as_content_owner/channel_spec.rb
@@ -55,6 +55,32 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'earnings can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+        let(:keys) { range.values }
+
+        specify 'with the :by option set to :range' do
+          earnings = channel.earnings range.merge by: :range
+          expect(earnings.size).to be 1
+          expect(earnings[:total]).to be_a Float
+        end
+      end
+
+      describe 'earnings can be grouped by day' do
+        let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
+        let(:keys) { range.values }
+
+        specify 'without a :by option (default)' do
+          earnings = channel.earnings range
+          expect(earnings.keys).to eq range.values
+        end
+
+        specify 'with the :by option set to :day' do
+          earnings = channel.earnings range.merge by: :day
+          expect(earnings.keys).to eq range.values
+        end
+      end
+
       describe 'views can be retrieved for a specific day' do
         context 'in which the channel was partnered' do
           let(:views) { channel.views_on 5.days.ago}
@@ -84,6 +110,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given end (:to option)' do
           expect(channel.views(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'views can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          views = channel.views range.merge by: :range
+          expect(views.size).to be 1
+          expect(views[:total]).to be_an Integer
         end
       end
 
@@ -200,6 +236,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'comments can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          comments = channel.comments range.merge by: :range
+          expect(comments.size).to be 1
+          expect(comments[:total]).to be_an Integer
+        end
+      end
+
       describe 'comments can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -244,6 +290,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given end (:to option)' do
           expect(channel.likes(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'likes can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          likes = channel.likes range.merge by: :range
+          expect(likes.size).to be 1
+          expect(likes[:total]).to be_an Integer
         end
       end
 
@@ -294,6 +350,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'dislikes can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          dislikes = channel.dislikes range.merge by: :range
+          expect(dislikes.size).to be 1
+          expect(dislikes[:total]).to be_an Integer
+        end
+      end
+
       describe 'dislikes can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -338,6 +404,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given end (:to option)' do
           expect(channel.shares(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'shares can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          shares = channel.shares range.merge by: :range
+          expect(shares.size).to be 1
+          expect(shares[:total]).to be_an Integer
         end
       end
 
@@ -388,6 +464,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'gained subscribers can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          subscribers_gained = channel.subscribers_gained range.merge by: :range
+          expect(subscribers_gained.size).to be 1
+          expect(subscribers_gained[:total]).to be_an Integer
+        end
+      end
+
       describe 'gained subscribers can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -432,6 +518,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given end (:to option)' do
           expect(channel.subscribers_lost(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'lost subscribers can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          subscribers_lost = channel.subscribers_lost range.merge by: :range
+          expect(subscribers_lost.size).to be 1
+          expect(subscribers_lost[:total]).to be_an Integer
         end
       end
 
@@ -482,6 +578,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'added favorites can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          favorites_added = channel.favorites_added range.merge by: :range
+          expect(favorites_added.size).to be 1
+          expect(favorites_added[:total]).to be_an Integer
+        end
+      end
+
       describe 'added favorites can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -529,6 +635,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'removed favorites can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          favorites_removed = channel.favorites_removed range.merge by: :range
+          expect(favorites_removed.size).to be 1
+          expect(favorites_removed[:total]).to be_an Integer
+        end
+      end
+
       describe 'removed favorites can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -573,6 +689,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given end (:to option)' do
           expect(channel.estimated_minutes_watched(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'estimated minutes watched can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          minutes = channel.estimated_minutes_watched range.merge by: :range
+          expect(minutes.size).to be 1
+          expect(minutes[:total]).to be_a Float
         end
       end
 
@@ -689,6 +815,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'average view duration can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          duration = channel.average_view_duration range.merge by: :range
+          expect(duration.size).to be 1
+          expect(duration[:total]).to be_a Float
+        end
+      end
+
       describe 'average view duration can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -733,6 +869,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given end (:to option)' do
           expect(channel.average_view_percentage(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'average view percentage can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          percentage = channel.average_view_percentage range.merge by: :range
+          expect(percentage.size).to be 1
+          expect(percentage[:total]).to be_a Float
         end
       end
 
@@ -783,6 +929,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'impressions can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          impressions = channel.impressions range.merge by: :range
+          expect(impressions.size).to be 1
+          expect(impressions[:total]).to be_an Integer
+        end
+      end
+
       describe 'impressions can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -830,6 +986,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'monetized playbacks can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          monetized_playbacks = channel.monetized_playbacks range.merge by: :range
+          expect(monetized_playbacks.size).to be 1
+          expect(monetized_playbacks[:total]).to be_an Integer
+        end
+      end
+
       describe 'monetized_playbacks can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -855,6 +1021,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given start (:from option) and a given end (:to option)' do
           expect(channel.annotation_clicks(from: date, to: date_to).keys.min).to eq date.to_date
+        end
+      end
+
+      describe 'annotation clicks can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          annotation_clicks = channel.annotation_clicks range.merge by: :range
+          expect(annotation_clicks.size).to be 1
+          expect(annotation_clicks[:total]).to be_an Integer
         end
       end
 
@@ -885,6 +1061,16 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'annotation click-through rate can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          rate = channel.annotation_click_through_rate range.merge by: :range
+          expect(rate.size).to be 1
+          expect(rate[:total]).to be_a Float
+        end
+      end
+
       describe 'annotation click-through rate can be grouped by day' do
         let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE'], until: Date.parse(ENV['YT_TEST_PARTNER_VIDEO_DATE']) + 5} }
 
@@ -909,6 +1095,16 @@ describe Yt::Channel, :partner do
 
         specify 'with a given start (:from option) and a given end (:to option)' do
           expect(channel.annotation_close_rate(from: date, to: date_to).keys.min).to eq date.to_date
+        end
+      end
+
+      describe 'annotation close rate can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          rate = channel.annotation_close_rate range.merge by: :range
+          expect(rate.size).to be 1
+          expect(rate[:total]).to be_a Float
         end
       end